BackEnd/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
google-generativeai==0.8.5